+++ CHANGELOG.md
@@ -18,6 +18,10 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 * Unlock the page scroll when the component is destroyed `bBottomSlide`
 * Now styles for dynamically loaded components with `loadModules` are included in the fat-html build `build/monic`
 
+#### :rocket: New Feature
+
+* Added the `unlockPageScrollOnDestroy` method `iLockPageScroll`
+
 ## v4.0.0-beta.93 (2024-04-19)
 
 #### :house: Internal